Implement an asset manager and a new menu navigation helper
Also rewrite parts of the button library for everything to work together.
The asset manager is currently barley used for anything and will be integrated
in further commits.

+++ td-racing-remake/core/src/com/mygdx/game/gamestate/states/resources/HighscoreButton.java
@@ -22,7 +22,7 @@ public class HighscoreButton implements Disposable {
 		this.name = name;
 		this.score = score;
 		this.button = new Sprite(HighscoreButton.texture);
-		this.button.setSize(HighscoreButton.texture.getWidth() / 2, HighscoreButton.texture.getHeight() / 2);
+		this.button.setSize((float) HighscoreButton.texture.getWidth() / 2, (float) HighscoreButton.texture.getHeight() / 2);
 		this.button.setPosition(xPosition - this.button.getWidth() / 2, yPosition - this.button.getHeight() / 2);
 		this.fontXNumber = xPosition - this.button.getWidth() / 2 + 20;
 		this.fontYNumber = yPosition + this.button.getHeight() / 5 * 2;
